requirements.txt
tkinter
websocket-client
requests
json
threading
time
collections
numpy
matplotlib
mpl_finance